specl: implement an alternate format, selected by -v option.
* specl (options): Accept -v or -verbose.
* src/specl.lua (run): Add new format parameter.  Assuming what
is passed conforms to the API, that formatter will be called as
the examples are executed.
(formatter): Move from here...
(report): ...to here.
(progress): New formatter that just prints dots as examples pass,
or F as they fail.
(formatter): Default formatter is the new progress formatter.
(M): Put the two formatters in the public interface.
* specl: Request the old report formatter in verbose mode.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>