matchers: diagnose passing expectations in pending examples.
* specs/specl_spec.yaml: Rearrange slightly to make room for a
new specification for unexpected passes in pending examples.
* README.md (Pending Examples): Document reporting of unexpected
passes with the built in formatters.
(Custom Formatters): Document changes to the expectations API
needed to support this.
* specl/colors.lua (colormap.warn, colormap.strong): New colors
for unexpected pass reporting.
* specl/formatter/progress.lua (expectations): Return unexpected
pass report strings, and display a `?` in the progress bar.
(footer): Report unexpected passes in the summaries.
* specl/formatter/report.lua (expectations, footer): Likewise,
except of course there is no progress bar with this formatter,
so display the unexpected pass report in the main output when
verbose is set.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>