specl: show description and offset line for example compile errors.
Previously, if the Lua code of any example failed, Specl would
dump the stack, usually containing a useless list of nested
run_context, run_examples etc etc.  We can already do a much
better job than that... libyaml needs to provide the support
for full and proper filewise line numbers though.
* specs/specl_spec.yaml (it reports line of error): Example to
demonstrate that the offset line of a compile failure is shown
on stderr.
* src/specl.lua (compile_example): Add a description parameter,
and use it as a prefix for displaying the line offset for the
failure, along with the Lua error.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>