specl: much improved optparse specs, prompting some bug fixes.
* specs/optparse_spec.yaml: Many, many more examples.
* specl/optparse.lua (normalise): Don't add a phantom '-' to the
end of split short options.
(finished): Add to unrecognised list options *after* the '--' that
got us here.
(opterr): Add a missing period to the end of passed error message
if necessary.
(parse): Make unrecognised list an instance variable so that it
can be augmented in `finished`.
(M.new): Capture and handle `--` first, as it is a substring of
long options and might match twice otherwise.
Accept a comma after a long option to allow multiple spellings of
a long option on the same line.
Consume only one `-` prefixed word per iteration to ensure the
handler patterns are compared against each one.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>