matchers: contain matcher now accepts std.objects.
Now that stdlib objects convert to tables by default (which in
turn prettyprint with ordered keys suitable for expectations to
compare against), should_contain can coerce an unknown object to
a table and search within that.
* specs/matchers_spec.yaml: Specify desired behaviour when passing
an object to the contain matcher.
* lib/specl/util.lua (type_check): Support "object" in the type
list, to denote a table that object.type recognises.
(M.totable): Re-export this from specl.std.
* lib/specl/matchers.lua (matchers.contain): Accept an object, and
coerce it to a table with std.table for comparison purposes.
* docs/specl.md (contain): Update.
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>