loader: report errors with macro expansion in Lua examples.
Although macro.substitute_tostring doesn't raise any errors when
the macro expansion is foiled, the errors it returns refer to
the LuaMacro internal stack. Consequently, we overwrite those
errors with something that helps point to the real cause of the
problem - a parser error during expect() expansion.
* specs/specl_spec.yaml: Specify error messages from common errors
with incorrect number of closing parens to expect macro.
* lib/specl/loader.lua (parser_mt:compile): Make macro expansion
a separate step, and report any resulting errors before moving on
to compilation of the result.
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>