configury: don't require help2man on build host.
As a general rule, configure should not generate distributed files.
In particular, we don't want configure to always generate a new
`bin/specl`, because that forces make to rebuild `doc/specl.1',
which requires help2man, which may not be installed on the build
host.
* configure.ac (AC_CONFIG_FILES): Remove bin/specl builder.
* local.mk ($(srcdir)/bin/specl): New rule to regenerate bin/specl
only when lib/specl.in has changed.
($(srcdir)/docs/specl.1): Calculate mkdir directory dynamically.
* specs/spec.mk ($(SPECL)): Always run distributed bin/specl from
$(srcdir).
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>