progress: don't try to animate a '>' pseudo-cursor in output.
Using ^H to backspace over the '>' character to simulate moving
it along the front of the output is unnecessary (the actual
cursor is already doing that), slow (there's a noticeable multi-
second pause on terminals that are capable of backspacing over
a carriage return to erase the last '>' on the previous line),
and error-prone (testing programs that switch to the alternate
screen between '>' and backspace often damages the display when
returning from the alternate screen).
* lib/specl/formatter/progress.lua (header): Remove. Adjust module
table.
(expectations): Don't output the '>' or try to erase it with a
subsequent backspace.
* specs/formatter/progress_spec.yaml: No need to match the old
invisible '>' and backspace characters with expect patterns.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>