environments: support requiring core modules in spec-files.
* specs/environment_spec.yaml (when requiring a module): Specify
behaviour when requiring a core module.
* lib/specl/runner.lua (initenv): A core module only has an entry
in `package.loaded`, but not in `package.preloaded`... and in any
case has no side-effects to capture since it's table is available
from _G in the outermost environment. So we can safely skip the
previously failing code for capturing side-effects when running
a module load function.
* NEWS: Update with note about bug-fix.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>